+++ auth-web/src/components/auth/create-account/AccountCreateBasic.vue
