+++ auth-web/src/components/auth/create-account/AccountTypeSelector.vue
