+++ auth-web/src/components/auth/create-account/non-bcsc/UploadAffidavitStep.vue
@@ -73,8 +73,8 @@ import { Address } from '@/models/address'
 import BcolLogin from '@/components/auth/BcolLogin.vue'
 import ConfirmCancelButton from '@/components/auth/common/ConfirmCancelButton.vue'
 import FileUploadPreview from '@/components/auth/common/FileUploadPreview.vue'
-import NotaryContactForm from '@/components/auth/NonBcscAccounts/NotaryContactForm.vue'
-import NotaryInformationForm from '@/components/auth/NonBcscAccounts/NotaryInformationForm.vue'
+import NotaryContactForm from '@/components/auth/create-account/non-bcsc/NotaryContactForm.vue'
+import NotaryInformationForm from '@/components/auth/create-account/non-bcsc/NotaryInformationForm.vue'
 import OrgModule from '@/store/modules/org'
 import { Organization } from '@/models/Organization'
 import Steppable from '@/components/auth/stepper/Steppable.vue'
