+++ auth-web/src/routes/router.ts
@@ -1,14 +1,14 @@
 import { Pages, Role, SessionStorageKeys } from '@/util/constants'
 import AcceptInviteLandingView from '@/views/auth/AcceptInviteLandingView.vue'
 import AcceptInviteView from '@/views/auth/AcceptInviteView.vue'
-import AccountChangeSuccessView from '@/views/auth/AccountChangeSuccessView.vue'
-import AccountChangeView from '@/views/auth/AccountChangeView.vue'
-import AccountCreationSuccessView from '@/views/auth/AccountCreationSuccessView.vue'
-import AccountInstructions from '@/components/auth/NonBcscAccounts/AccountInstructions.vue'
+import AccountChangeSuccessView from '@/views/auth/create-account/AccountChangeSuccessView.vue'
+import AccountChangeView from '@/views/auth/create-account/AccountChangeView.vue'
+import AccountCreationSuccessView from '@/views/auth/create-account/AccountCreationSuccessView.vue'
+import AccountInstructions from '@/components/auth/create-account/non-bcsc/AccountInstructions.vue'
 import AccountLoginOptionsChooser from '@/views/auth/AccountLoginOptionsChooser.vue'
 import AccountLoginOptionsInfo from '@/views/auth/AccountLoginOptionsInfo.vue'
-import AccountSetupView from '@/views/auth/AccountSetupView.vue'
-import AffidavitDownload from '@/components/auth/NonBcscAccounts/AffidavitDownload.vue'
+import AccountSetupView from '@/views/auth/create-account/AccountSetupView.vue'
+import AffidavitDownload from '@/components/auth/create-account/non-bcsc/AffidavitDownload.vue'
 import AuthenticationOptionsView from '@/views/auth/AuthenticationOptionsView.vue'
 import BusinessProfileView from '@/views/auth/BusinessProfileView.vue'
 import ChooseAuthMethodView from '@/views/auth/ChooseAuthMethodView.vue'
@@ -23,9 +23,9 @@ import HomeView from '@/views/auth/HomeView.vue'
 import IncorpOrRegisterView from '@/views/auth/IncorpOrRegisterView.vue'
 import LeaveTeamLandingView from '@/views/auth/LeaveTeamLandingView.vue'
 import MaintainBusinessView from '@/views/auth/MaintainBusinessView.vue'
-import NonBcscAccountCreationSuccessView from '@/views/auth/NonBcscAccountCreationSuccessView.vue'
-import NonBcscAccountSetupView from '@/views/auth/NonBcscAccountSetupView.vue'
-import NonBcscInfoView from '@/views/auth/NonBcscAccountView.vue'
+import NonBcscAccountCreationSuccessView from '@/views/auth/create-account/non-bcsc/NonBcscAccountCreationSuccessView.vue'
+import NonBcscAccountSetupView from '@/views/auth/create-account/non-bcsc/NonBcscAccountSetupView.vue'
+import NonBcscInfoView from '@/views/auth/create-account/non-bcsc/NonBcscInfoView.vue'
 import PageNotFound from '@/views/auth/PageNotFound.vue'
 import PaymentReturnView from '@/views/pay/PaymentReturnView.vue'
 import PaymentView from '@/views/pay/PaymentView.vue'
