+++ auth-web/src/views/auth/DuplicateTeamWarningView.vue
@@ -19,7 +19,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
 import { mapActions, mapState } from 'vuex'
-import CreateAccountInfoForm from '@/components/auth/CreateAccount/CreateAccountInfoForm.vue'
+import CreateAccountInfoForm from '@/components/auth/create-account/CreateAccountInfoForm.vue'
 import { Organization } from '@/models/Organization'
 import { getModule } from 'vuex-module-decorators'
