+++ auth-web/src/views/auth/create-account/AccountCreationSuccessView.vue
