+++ auth-web/src/views/auth/create-account/non-bcsc/NonBcscAccountSetupView.vue
@@ -14,11 +14,11 @@
 import { Component, Vue } from 'vue-property-decorator'
 import Stepper, { StepConfiguration } from '@/components/auth/stepper/Stepper.vue'
 import { AccessType } from '@/util/constants'
-import AccountCreateBasic from '@/components/auth/CreateAccount/AccountCreateBasic.vue'
-import AccountCreatePremium from '@/components/auth/CreateAccount/AccountCreatePremium.vue'
-import AccountTypeSelector from '@/components/auth/CreateAccount/AccountTypeSelector.vue'
-import CreateAccountInfoForm from '@/components/auth/CreateAccount/CreateAccountInfoForm.vue'
-import UploadAffidavitStep from '@/components/auth/NonBcscAccounts/UploadAffidavitStep.vue'
+import AccountCreateBasic from '@/components/auth/create-account/AccountCreateBasic.vue'
+import AccountCreatePremium from '@/components/auth/create-account/AccountCreatePremium.vue'
+import AccountTypeSelector from '@/components/auth/create-account/AccountTypeSelector.vue'
+import CreateAccountInfoForm from '@/components/auth/create-account/CreateAccountInfoForm.vue'
+import UploadAffidavitStep from '@/components/auth/create-account/non-bcsc/UploadAffidavitStep.vue'
 import UserProfileForm from '@/components/auth/UserProfileForm.vue'
 
 @Component({
