+++ auth-web/src/views/auth/create-account/non-bcsc/NonBcscInfoView.vue
